스컬 콤보 공격 수정, 웨어울프 스킬 4종 추가, 스위치 기능 수정 및 추가

+++ Assets/3.Script/Etc/Resolution.cs
@@ -21,7 +21,7 @@ public class Resolution : MonoBehaviour
 
     public void SetAspect()
     {
-        float curResolution = (float)Screen.width / (float)Screen.height;
+        float curResolution = (float)Screen.width / Screen.height;
 
         Rect viewportRect = cam.rect;
